+++ app/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.0.0.3"
-__asset_rev__ = "20260914g"
+__version__ = "0.0.0.4"
+__asset_rev__ = "20260914i"
 __author__ = "Adam Stone"